fix(visualize): remove scienceplots integration
The `_style.py` module relied on `scienceplots` plus internal matplotlib
APIs (`plt.style.core`, `plt.style.library['notebook']`) that newer
matplotlib no longer exposes, causing
`AttributeError: module 'matplotlib.style' has no attribute 'core'` and
breaking CI on py3.13 across all platforms.

The module was never imported by the package (not in any `__init__`,
not in `__all__`, and `scienceplots` was never a declared dependency),
so it is removed along with its `TestStyleModule` tests. The unrelated
`get_figure` tests that were bundled into `_style_test.py` are preserved
in a new `_figures_test.py`.
